% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2num.R
\name{bin2num}
\alias{bin2num}
\title{binary to numeric}
\usage{
bin2num(dat, len, na = NA, clutter = NA, RX = FALSE)
}
\arguments{
\item{dat}{Binary data returned by \code{\link{readBin}}}

\item{len}{Length of data.}

\item{na}{Value to be set for missing data (bit 14). DEFAULT: NA}

\item{clutter}{Value to be set for clutter data (bit 16). DEFAULT: NA}

\item{RX}{Logical: call rx routine? DEFAULT: FALSE}
}
\value{
numerical vector
}
\description{
Call FORTRAN routines
}
\seealso{
\code{\link{readRadarFile}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May + Oct 2019
}
\keyword{file}
