% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHeader.R
\name{readHeader}
\alias{readHeader}
\title{Read header part of binary DWD files}
\usage{
readHeader(file)
}
\arguments{
\item{file}{Name of a single binary file}
}
\value{
List with original string, nchar, derived information
}
\description{
Read and process header of binary radar files
}
\examples{
# See readRadarFile

}
\seealso{
Used in \code{\link{readRadarFile}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2020
}
\keyword{binary}
\keyword{file}
